refactor(ihv): extract provider card and memory offload components
- Extract ProviderCardGrid component to separate grid rendering logic from panel
- Create MemoryOffloadControls component for hybrid offload UI and state management
- Extract LocalFileUpload component for recipe input handling
- Add RecipeJsonEditor component for direct JSON recipe editing
- Add RecipeCatalogBrowser component for browsing available recipes
- Add GitHubRecipeSync component for syncing recipes from GitHub
- Simplify IHVIntegrationPanel by delegating to specialized subcomponents
- Update InputEnvironmentPanel to use new input components
- Improve code organization and maintainability by reducing component complexity
