+++ tsconfig.json
@@ -4,7 +4,11 @@
     "experimentalDecorators": true,
     "useDefineForClassFields": false,
     "module": "ESNext",
-    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "lib": [
+      "ES2022",
+      "DOM",
+      "DOM.Iterable"
+    ],
     "skipLibCheck": true,
     "moduleResolution": "bundler",
     "resolveJsonModule": true,
@@ -12,12 +16,20 @@
     "moduleDetection": "force",
     "allowJs": true,
     "jsx": "react-jsx",
-    "types": ["@webgpu/types"],
+    "types": [
+      "@webgpu/types"
+    ],
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": [
+        "./src/*"
+      ]
     },
     "allowImportingTsExtensions": true,
     "noEmit": true
   },
-  "exclude": ["src-tauri/target", "dist", "node_modules"]
+  "exclude": [
+    "src-tauri/target",
+    "dist",
+    "node_modules"
+  ]
 }
